requirements.txt
rich-argparse
argcomplete
google_api_python_client>=2.125.0
google_auth_oauthlib>=1.2.0
grpcio>=1.62.0
grpcio-tools>=1.62.0
inquirer>=3.3.0
litellm>=1.40.0
pexpect>=4.8.0
protobuf>=6.31.1,<7.0.0
pycryptodome>=3.18.0
pyfzf>=0.3.1
PyYAML>=6.0.1
rich>=13.7.1